build: src/app-only; chat gate bypass in preview
