--- app/centro-de-ajuda/page.tsx
@@ -1,13 +1,0 @@
-export const metadata = {
-  title: "Centro de Ajuda — CRSET",
-  description: "Perguntas frequentes e suporte",
-  alternates: { canonical: "/centro-de-ajuda" },
-};
-export default function Page() {
-  return (
-    <main className="container mx-auto max-w-5xl px-4 py-10 space-y-4">
-      <h1 className="text-2xl font-semibold">Centro de Ajuda</h1>
-      <p className="text-muted-foreground">Conteúdo em breve.</p>
-    </main>
-  );
-}
