+++ src/app/.well-known/appspecific/com.chrome.devtools.json/route.ts
