+++ src/app/api/agi/chat/route.ts
