+++ src/app/api/agi/metrics/route.ts
