+++ src/app/api/agi/status/route.ts
