+++ src/app/api/agi/stream/route.ts
