+++ src/app/api/cron/health/route.ts
