+++ src/app/api/debug/sentry/route.ts
