+++ src/app/api/status/route.ts
