+++ src/app/components/ChatWidget.tsx
