+++ src/app/global-error.tsx
@@ -1,23 +1,23 @@
-"use client";
+"use client"
 
-import * as Sentry from "@sentry/nextjs";
-import NextError from "next/error";
-import { useEffect } from "react";
-
-export default function GlobalError({ error }: { error: Error & { digest?: string } }) {
-  useEffect(() => {
-    Sentry.captureException(error);
-  }, [error]);
-
+export default function GlobalError({
+  error,
+  reset,
+}: {
+  error: Error & { digest?: string }
+  reset: () => void
+}) {
+  console.error("[GlobalError]", error)
   return (
     <html>
       <body>
-        {/* `NextError` is the default Next.js error page component. Its type
-        definition requires a `statusCode` prop. However, since the App Router
-        does not expose status codes for errors, we simply pass 0 to render a
-        generic error message. */}
-        <NextError statusCode={0} />
+        <div style={{ padding: 16 }}>
+          <h2>Algo correu mal.</h2>
+          <button onClick={() => reset()} style={{ marginTop: 8 }}>
+            Tentar de novo
+          </button>
+        </div>
       </body>
     </html>
-  );
+  )
 }
